@@ -1,90 +1,99 @@
 {
-  "name": "pdp-2022",
-  "private": true,
-  "version": "0.0.1",
-  "description": "React based project for pdp-2022",
-  "repository": "https://github.com/blackskip-notOk/react-stack-pdp-21-22",
-  "author": "Ilya Babanov <ilbabanov@gmail.com>",
-  "license": "MIT",
-  "scripts": {
-    "start": "vite --open",
-    "build": "tsc && vite build",
-    "preview": "vite preview",
-    "prepare": "husky install",
-    "lint": "eslint src/**/*.ts src/**/*.tsx --ignore-pattern node_modules/",
-    "test": "jest",
-    "vitest": "vitest"
-  },
-  "dependencies": {
-    "@ant-design/icons": "^4.7.0",
-    "@emotion/react": "^11.8.2",
-    "@emotion/styled": "^11.8.1",
-    "@hookform/resolvers": "^2.8.8",
-    "@mui/icons-material": "^5.5.1",
-    "@mui/lab": "^5.0.0-alpha.74",
-    "@mui/material": "^5.5.2",
-    "@mui/styled-engine-sc": "^5.5.2",
-    "@types/ramda": "^0.27.64",
-    "antd": "^4.18.7",
-    "axios": "^0.25.0",
-    "effector": "^22.2.0",
-    "effector-localstorage": "^0.5.3",
-    "effector-react": "^22.0.6",
-    "effector-storage": "^5.0.0",
-    "effector-utils": "^0.1.0",
-    "node-notifier": "^10.0.1",
-    "path": "^0.12.7",
-    "postcss": "^8.4.6",
-    "ramda": "^0.28.0",
-    "react": "^17.0.2",
-    "react-dom": "^17.0.2",
-    "react-hook-form": "^7.28.1",
-    "react-query": "^3.34.13",
-    "react-router-dom": "^6.2.2",
-    "react-spinners": "^0.11.0",
-    "rxjs": "^7.5.5",
-    "yup": "^0.32.11"
-  },
-  "devDependencies": {
-    "@reactivex/rxjs": "^6.6.7",
-    "@testing-library/dom": "^8.11.3",
-    "@testing-library/jest-dom": "^5.16.2",
-    "@testing-library/react": "^12.1.3",
-    "@testing-library/user-event": "^13.5.0",
-    "@types/jest": "^27.4.0",
-    "@types/react": "^17.0.33",
-    "@types/react-dom": "^17.0.10",
-    "@types/react-test-renderer": "^17.0.1",
-    "@vitejs/plugin-react": "^1.0.7",
-    "effector-logger": "^0.13.4",
-    "eslint": "^8.8.0",
-    "eslint-config-prettier": "^8.3.0",
-    "eslint-import-resolver-alias": "^1.1.2",
-    "eslint-plugin-effector": "^0.6.0",
-    "eslint-plugin-import": "^2.25.4",
-    "eslint-plugin-jsx-a11y": "^6.5.1",
-    "eslint-plugin-prettier": "^4.0.0",
-    "eslint-plugin-react": "^7.28.0",
-    "eslint-plugin-react-hooks": "^4.3.0",
-    "husky": "^7.0.0",
-    "identity-obj-proxy": "^3.0.0",
-    "jest": "^27.5.1",
-    "jest-circus": "^27.5.1",
-    "jest-watch-typeahead": "^1.0.0",
-    "less": "^4.1.2",
-    "msw": "^0.39.0",
-    "postcss-load-config": "^3.1.3",
-    "pre-commit": "^1.2.2",
-    "prettier": "2.5.1",
-    "ts-jest": "^27.1.3",
-    "ts-node": "^10.5.0",
-    "typescript": "^4.5.4",
-    "vite": "^2.8.4",
-    "vite-aliases": "^0.8.7",
-    "vitest": "^0.7.7"
-  },
-  "pre-commit": [
-    "lint",
-    "test"
-  ]
+	"name": "pdp-2022",
+	"private": true,
+	"version": "0.0.1",
+	"description": "React based project for pdp-2022",
+	"repository": "https://github.com/blackskip-notOk/react-stack-pdp-21-22",
+	"author": "Ilya Babanov <ilbabanov@gmail.com>",
+	"license": "MIT",
+	"scripts": {
+		"start": "vite --open",
+		"build": "tsc && vite build",
+		"preview": "vite preview",
+		"prepare": "husky install",
+		"test": "jest --bail --runInBand --coverage=false",
+		"lint": "eslint --debug src/",
+		"lint:write": "eslint --debug src/ --fix",
+		"prettier": "prettier --write src/**/*.{ts,tsx}",
+		"postinstall": "husky install",
+		"test:commit": "jest --bail --lastCommit --logHeapUsage --coverage=true"
+	},
+	"dependencies": {
+		"@emotion/react": "^11.8.2",
+		"@emotion/styled": "^11.8.1",
+		"@hookform/resolvers": "^2.8.8",
+		"@mui/icons-material": "^5.5.1",
+		"@mui/lab": "^5.0.0-alpha.75",
+		"@mui/material": "^5.5.3",
+		"@mui/styled-engine-sc": "^5.5.2",
+		"@types/ramda": "^0.28.7",
+		"axios": "^0.26.1",
+		"effector": "^22.2.0",
+		"effector-localstorage": "^0.5.3",
+		"effector-react": "^22.0.6",
+		"effector-storage": "^5.0.0",
+		"effector-utils": "^0.1.0",
+		"node-notifier": "^10.0.1",
+		"path": "^0.12.7",
+		"postcss": "^8.4.12",
+		"ramda": "^0.28.0",
+		"react": "^18.0.0",
+		"react-dom": "^18.0.0",
+		"react-hook-form": "^7.29.0",
+		"react-router-dom": "^6.3.0",
+		"react-spinners": "^0.11.0",
+		"rxjs": "^7.5.5",
+		"yup": "^0.32.11"
+	},
+	"devDependencies": {
+		"@reactivex/rxjs": "^6.6.7",
+		"@testing-library/jest-dom": "^5.16.3",
+		"@testing-library/react": "^13.0.0",
+		"@testing-library/user-event": "^14.0.4",
+		"@types/jest": "^27.4.1",
+		"@types/react": "^17.0.43",
+		"@types/react-dom": "^17.0.14",
+		"@typescript-eslint/eslint-plugin": "^5.17.0",
+		"@typescript-eslint/parser": "^5.17.0",
+		"@vitejs/plugin-react": "^1.3.0",
+		"effector-logger": "^0.13.4",
+		"eslint": "^8.12.0",
+		"eslint-config-prettier": "^8.5.0",
+		"eslint-import-resolver-alias": "^1.1.2",
+		"eslint-plugin-effector": "^0.6.0",
+		"eslint-plugin-import": "^2.25.4",
+		"eslint-plugin-jsx-a11y": "^6.5.1",
+		"eslint-plugin-prettier": "^4.0.0",
+		"eslint-plugin-react": "^7.29.4",
+		"eslint-plugin-react-hooks": "^4.4.0",
+		"eslint-plugin-simple-import-sort": "^7.0.0",
+		"husky": "^7.0.0",
+		"identity-obj-proxy": "^3.0.0",
+		"jest": "^27.5.1",
+		"jest-circus": "^27.5.1",
+		"jest-watch-typeahead": "^1.0.0",
+		"less": "^4.1.2",
+		"msw": "^0.39.2",
+		"postcss-load-config": "^3.1.4",
+		"pre-commit": "^1.2.2",
+		"prettier": "2.6.2",
+		"pretty-quick": "^3.1.3",
+		"ts-jest": "^27.1.4",
+		"ts-node": "^10.7.0",
+		"typescript": "^4.6.3",
+		"vite": "^2.9.1",
+		"vite-aliases": "^0.9.2"
+	},
+	"husky": {
+		"hooks": {
+			"pre-commit": "lint-staged",
+			"pre-push": "test"
+		}
+	},
+	"lint-staged": {
+		"src/**/*.{ts,tsx}": [
+			"yarn run lint:write",
+			"yarn run prettier"
+		]
+	}
 }
