@@ -1,11 +1,17 @@
-import ReactDOM from 'react-dom'
-import { App } from './App'
+import { createRoot } from 'react-dom/client';
+import { App } from './App';
 import { BrowserRouter } from 'react-router-dom';
 import './models/init';
+import { APP_MESSAGES } from './constants/serverMessages';
 
-ReactDOM.render(
+const container = document.getElementById('root');
+
+if (!container) throw new Error(`${APP_MESSAGES.NOT_ROOT}`);
+
+const root = createRoot(container);
+
+root.render(
 	<BrowserRouter>
 		<App />
 	</BrowserRouter>,
-	document.getElementById('root')
 );
