@@ -1,12 +1,28 @@
 module.exports = {
-	env: {
-		browser: true,
-		es2021: true,
+	parser: '@typescript-eslint/parser',
+	parserOptions: {
+		// project: ['./tsconfig.json'],
+		ecmaFeatures: {
+			jsx: true,
+		},
+		ecmaVersion: 'latest',
+		sourceType: 'module',
 	},
 	settings: {
 		react: {
 			version: 'detect',
 		},
+		// 'import/resolver': {
+		// 	node: {
+		// 	  paths: ['src'],
+		// 	  extensions: ['.js', '.jsx', '.ts', '.tsx'],
+		// 	},
+		//   },
+	},
+	env: {
+		browser: true,
+		es2021: true,
+		node: true,
 	},
 	extends: [
 		'eslint:recommended',
@@ -15,36 +31,22 @@ module.exports = {
 		'plugin:react-hooks/recommended',
 		'plugin:jsx-a11y/recommended',
 		'plugin:@typescript-eslint/recommended',
-		'prettier',
+		'plugin:prettier/recommended',
 	],
-	parser: '@typescript-eslint/parser',
-	parserOptions: {
-		project: ['./tsconfig.json'],
-		ecmaFeatures: {
-			jsx: true,
-		},
-		ecmaVersion: 'latest',
-		sourceType: 'module',
-	},
-	plugins: [
-		'react',
-		'react-hooks',
-		'jsx-a11y',
-		'@typescript-eslint',
-		'prettier',
-		"effector",
-	],
+	plugins: ['react', 'react-hooks', 'jsx-a11y', '@typescript-eslint', 'prettier', 'effector'],
 	rules: {
+		'linebreak-style': 0,
+		'no-plusplus': 0,
+		'no-param-reassign': 2,
+		'react/require-default-props': 0,
+		'react-hooks/exhaustive-deps': 0,
 		'prettier/prettier': 0,
 		'arrow-body-style': 0,
 		'prefer-arrow-callback': 0,
-		camelcase: [
-			1,
-			{ ignoreDestructuring: true, ignoreImports: true, ignoreGlobals: true },
-		],
+		camelcase: [1, { ignoreDestructuring: true, ignoreImports: true, ignoreGlobals: true }],
 		'react-hooks/rules-of-hooks': 2,
-		'react-hooks/exhaustive-deps': 1,
 		'react/react-in-jsx-scope': 0,
+		'max-len': [1, { code: 100, ignoreUrls: true }],
 	},
 	ignorePatterns: ['src/**/*.test.ts'],
 };
