@@ -3,6 +3,7 @@ export default {
 		name: 'REACT_APP',
 		color: 'magenta',
 	},
+	roots: ['<rootDir>/src'],
 	testEnvironment: 'jsdom',
 	transform: {
 		'^.+\\.tsx?$': 'ts-jest',
@@ -10,15 +11,20 @@ export default {
 	moduleNameMapper: {
 		'\\.(gif|ttf|eot|svg|png)$': '<rootDir>/test/__mocks__/fileMock.js',
 		'\\.(css|less|sass|scss)$': 'identity-obj-proxy',
-		'^@/(.*)$': '<rootDir>/src/$1',
 	},
 	setupFilesAfterEnv: ['<rootDir>/jest.setup.ts'],
-	collectCoverage: true,
-	collectCoverageFrom: ['**/*.{ts, tsx, js,jsx}', '!**/node_modules/**'],
+	collectCoverage: false,
+	collectCoverageFrom: [
+		'!**/*.config.{ts, js}',
+		'!**/*.d.ts',
+		'!**/src/mocks/**',
+	],
+	testRegex: '(/__tests__/.*|(\\.|/)(test|spec))\\.tsx?$',
 	coverageDirectory: '<rootDir>/coverage',
 	coveragePathIgnorePatterns: ['<rootDir>/.husky/', '<rootDir>/node_modules/'],
 	errorOnDeprecated: true,
-	moduleFileExtensions: ["tsx", "ts", "js", "jsx", "json", "node"],
+	moduleFileExtensions: ['tsx', 'ts', 'js', 'jsx', 'json', 'node'],
 	notify: true,
 	notifyMode: 'failure-change',
+	bail: true,
 };
