@@ -1,13 +1,16 @@
 import '@testing-library/jest-dom/extend-expect';
 import { server } from './src/mocks/server';
-import { QueryCache } from 'react-query'
 
-const queryCache = new QueryCache()
-afterEach(() => {
-  queryCache.clear()
-});
-
-beforeAll(() => server.listen());
+beforeAll(() =>
+	server.listen({
+		onUnhandledRequest: ({ headers }, print) => {
+			if (headers.get('User-Agent') === 'supertest') {
+				return;
+			}
+			print.error();
+		},
+	}),
+);
 
 afterEach(() => server.resetHandlers());
 
