@@ -3,7 +3,11 @@ import { render } from '@testing-library/react';
 import { BrowserRouter } from 'react-router-dom';
 
 test('Renders main page correctly', () => {
-    render(<BrowserRouter><App /></BrowserRouter>);
+	render(
+		<BrowserRouter>
+			<App />
+		</BrowserRouter>,
+	);
 
-  expect(true).toBeTruthy();
+	expect(true).toBeTruthy();
 });
