@@ -2,12 +2,10 @@ import { FC } from 'react';
 import { Posts } from '../Post/Posts';
 
 export const Home: FC = () => {
-
-
-
-
-	return <>
-	<div>Home</div>
-	<Posts />
-	</>;
-}
+	return (
+		<>
+			<div>Home</div>
+			<Posts />
+		</>
+	);
+};
