@@ -6,11 +6,12 @@ import styles from './Login.module.less';
 import { $auth } from '../../models/auth/index';
 
 export const Login: FC = () => {
+	const { message, isAuth } = useStore($auth);
 
-	const {message, isAuth} = useStore($auth);
-
-	const localMessage = message === SERVER_MESSAGES.NOT_AUTHORIZED
-	? SERVER_MESSAGES_DESCRIPTIONS.notAutorized : message;
+	const localMessage =
+		message === SERVER_MESSAGES.NOT_AUTHORIZED
+			? SERVER_MESSAGES_DESCRIPTIONS.notAutorized
+			: message;
 
 	return (
 		<div className={styles.loginContainer}>
