@@ -1,12 +1,14 @@
-import { FC } from 'react'
-import { NavLink } from 'react-router-dom'
-import { NAVLINKS } from '../../constants/routerConstants'
+import { FC } from 'react';
+import { NavLink } from 'react-router-dom';
+import { NAVLINKS } from '../../constants/routerConstants';
 
 export const NavBar: FC = () => {
 	return (
 		<>
-	<div><NavLink to={`${NAVLINKS.PROFILE}`}>Profile</NavLink></div>
-	<NavLink to={`${NAVLINKS.LOGIN}`}>Login</NavLink>
-	</>
-	)
-}
+			<div>
+				<NavLink to={`${NAVLINKS.PROFILE}`}>Profile</NavLink>
+			</div>
+			<NavLink to={`${NAVLINKS.LOGIN}`}>Login</NavLink>
+		</>
+	);
+};
