@@ -1,12 +1,12 @@
-import { FC } from "react";
+import { FC } from 'react';
 
 export const NotFound: FC = () => {
-    return (
-        <div>
-            <img src="https://i.imgur.com/qIufhof.png" alt="not found page" />
-            <div>
-                <h3>This page could not be found</h3>
-            </div>
-        </div >
-    )
-}
+	return (
+		<div>
+			<img src='https://i.imgur.com/qIufhof.png' alt='not found page' />
+			<div>
+				<h3>This page could not be found</h3>
+			</div>
+		</div>
+	);
+};
