@@ -1,11 +1,11 @@
 export const RESPONSE_STATUSES = {
-    success: 200,
-    clientError: 404,
-    serverError: 500,
+	success: 200,
+	clientError: 404,
+	serverError: 500,
 };
 
 export const RESULT_CODES = {
-    success: 0,
-    error: 1,
-    secure: 10,
+	success: 0,
+	error: 1,
+	secure: 10,
 };
