@@ -1,26 +1,26 @@
-import { BaseResponse } from "../types";
+import { BaseResponse } from '../types';
 
 interface AuthData {
-    id: number,
-    email: string,
-    login: string
-};
+	id: number;
+	email: string;
+	login: string;
+}
 
 export interface Auth extends BaseResponse {
-    data: AuthData,
-};
+	data: AuthData;
+}
 
 export interface AuthResponse {
-    status: number,
-    authInfo: Auth
-};
+	status: number;
+	authInfo: Auth;
+}
 
 export interface AuthState {
-    isAuth: boolean,
-    message: string
-};
+	isAuth: boolean;
+	message: string;
+}
 
 export interface Owner {
-    isOwner: boolean,
-    ownerId: number | null,
-};
+	isOwner: boolean;
+	ownerId?: number;
+}
