@@ -1,5 +1,5 @@
 export interface BaseResponse {
-    resultCode: number,
-    messages: Array<string>,
-    fieldsErrors: Array<string>
-};
+	resultCode: number;
+	messages: Array<string>;
+	fieldsErrors: Array<string>;
+}
